Refactor by moving ClassInterval to LL folder and delete some deadcode
