+++ daedalus.cabal
@@ -60,8 +60,8 @@ library
                        Daedalus.ParserGen.Utils,
                        Daedalus.ParserGen.Cfg,
                        Daedalus.ParserGen.Compile,
-                       Daedalus.ParserGen.ClassInterval,
                        Daedalus.ParserGen.LL,
+                       Daedalus.ParserGen.LL.ClassInterval,
                        Daedalus.ParserGen.LL.Result,
                        Daedalus.ParserGen.LL.CfgDet,
                        Daedalus.ParserGen.LL.Closure,
